Move some helper methods to parent job
